[ANDROSDK-779] Extract CallableObservableTransformer; test SystemInfoCall as Observable
