+++ core/src/main/java/org/hisp/dhis/android/core/arch/api/executors/CallableObservableTransformer.java
@@ -25,30 +25,29 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.systeminfo;
 
-import org.hisp.dhis.android.core.arch.api.executors.CallableObservableTransformer;
-import org.hisp.dhis.android.core.arch.repositories.children.ChildrenAppender;
-import org.hisp.dhis.android.core.arch.repositories.object.ReadOnlyFirstObjectWithDownloadRepositoryImpl;
-import org.hisp.dhis.android.core.arch.repositories.scope.RepositoryScope;
-import org.hisp.dhis.android.core.common.ObjectWithoutUidStore;
+package org.hisp.dhis.android.core.arch.api.executors;
 
-import java.util.Map;
+import java.util.concurrent.Callable;
 
-import javax.inject.Inject;
+import io.reactivex.Single;
 
-import dagger.Reusable;
+public final class CallableObservableTransformer {
 
-@Reusable
-public final class SystemInfoObjectRepository
-        extends ReadOnlyFirstObjectWithDownloadRepositoryImpl<SystemInfo, SystemInfoObjectRepository> {
+    CallableObservableTransformer() {
+    }
 
-    @Inject
-    SystemInfoObjectRepository(ObjectWithoutUidStore<SystemInfo> store,
-                               Map<String, ChildrenAppender<SystemInfo>> childrenAppenders,
-                               RepositoryScope scope,
-                               SystemInfoCall systemInfoCall) {
-        super(store, childrenAppenders, scope, CallableObservableTransformer.toCallable(systemInfoCall.asObservable()),
-                cs -> new SystemInfoObjectRepository(store, childrenAppenders, cs, systemInfoCall));
+    public static <O> Callable<O> toCallable(Single<O> single) {
+        return () -> {
+            try {
+                return single.blockingGet();
+            } catch (Exception e) {
+                if (e.getCause() instanceof Exception) {
+                    throw (Exception) e.getCause();
+                } else {
+                    throw new RuntimeException(e.getCause());
+                }
+            }
+        };
     }
 }
